Create soft delte functionality

+++ Backend/officecallendar.backend/Controllers/MessagesController.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using Officecalendar.Backend.Models;
 using OfficeCalendar.Backend.DTOs;
 using OfficeCalendar.Backend.Services;
@@ -19,7 +20,7 @@ public class MessagesController : ControllerBase
     }
 
     [HttpGet]
-    public ActionResult<Message> GetMessage([FromQuery] Guid messageid)
+    public ActionResult<MessageGetDto> GetMessage([FromQuery] Guid messageid)
     {
         var message = _messageService.GetByGuid(messageid);
         if (message == null)
@@ -29,11 +30,42 @@ public class MessagesController : ControllerBase
                 message = $"Message with id {messageid} not found"
             });
 
-        return Ok(message);
+        if (message.sender_username != User.Identity.Name && !message.MessageReceivers.Any(u => u.username == User.Identity.Name))
+        {
+            return Forbid();
+        }
+
+        var dto = new MessageGetDto
+        {
+            sender_username = message.sender_username,
+            title = message.title,
+            desc = message.desc,
+            receivers = message.MessageReceivers.Select(r => r.username).ToList(),
+            referenced_event_id = message.referenced_event_id,
+            creation_date = message.creation_date,
+            last_edited_date = message.last_edited_date
+        };
+
+        return Ok(dto);
+    }
+
+    [HttpGet("me")]
+    public ActionResult<MessageGetDto[]> GetMessageForUser()
+    {
+        var messages = _messageService.GetMessagesForUser(User.Identity.Name);
+
+        if (messages.Length == 0)
+            return NotFound(new
+            {
+                statuscode = 404,
+                message = $"No messages for {User.Identity.Name} found"
+            });
+
+        return Ok(messages);
     }
 
     [HttpPost]
-    public ActionResult<Message> CreateMessage(MessagePostDto dto)
+    public ActionResult<MessageGetDto> CreateMessage(MessagePostDto dto)
     {
         if (!ModelState.IsValid)
             return BadRequest(
@@ -43,8 +75,28 @@ public class MessagesController : ControllerBase
                     message = ModelState
                 });
 
-        var message = _messageService.PostMessage(dto, User.Identity.Name);
-        return CreatedAtAction(nameof(GetMessage), new { messageid = message.id }, message);
+        var response = _messageService.PostMessage(dto, User.Identity.Name);
+        if (!response.success)
+        {
+            return NotFound(new
+            {
+                statuscode = 404,
+                message = response.error
+            });
+        }
+
+        var createdDto = new MessageGetDto
+        {
+            sender_username = response.message.sender_username,
+            title = response.message.title,
+            desc = response.message.desc,
+            receivers = response.message.MessageReceivers.Select(r => r.username).ToList(),
+            referenced_event_id = response.message.referenced_event_id,
+            creation_date = response.message.creation_date,
+            last_edited_date = response.message.last_edited_date
+        };
+
+        return CreatedAtAction(nameof(GetMessage), new { messageid = response.message.id }, createdDto);
     }
 
     [HttpDelete]
@@ -58,13 +110,18 @@ public class MessagesController : ControllerBase
                 message = $"Message with id {messageid} not found"
             });
 
+        if (message.sender_username != User.Identity.Name)
+        {
+            return Forbid();
+        }
+
         _messageService.DeleteMessage(message);
 
         return NoContent();
     }
 
     [HttpPut]
-    public ActionResult<Message> UpdateMessage([FromQuery] Guid messageid, MessagePutDto dto)
+    public ActionResult<MessageGetDto> UpdateMessage([FromQuery] Guid messageid, MessagePutDto dto)
     {
         var message = _messageService.GetByGuid(messageid);
         if (message == null)
@@ -74,7 +131,24 @@ public class MessagesController : ControllerBase
                 message = $"Message with id {messageid} not found"
             });
 
+        if (message.sender_username != User.Identity.Name)
+        {
+            return Forbid();
+        }
+
         _messageService.UpdateMessage(message, dto);
-        return Ok(message);
+
+        var updatedDto = new MessageGetDto
+        {
+            sender_username = message.sender_username,
+            title = message.title,
+            desc = message.desc,
+            receivers = message.MessageReceivers.Select(r => r.username).ToList(),
+            referenced_event_id = message.referenced_event_id,
+            creation_date = message.creation_date,
+            last_edited_date = message.last_edited_date
+        };
+
+        return Ok(updatedDto);
     }
 }
